+++ public/javascripts/bootstrap_server_templating.js
@@ -1,5 +1,6 @@
 $(document).ready(function(){
-    var socket = io.connect('127.0.0.1',{ port: 3000 });
+    var socket = io.connect('127.0.0.1',{ port: 8080 });
+    //var socket = io.connect('https://127.0.0.1:8443/', {secure: true, port: '8443'});
     node_plates.initialise(socket, page);
     node_plates.server_template();