+++ server.js
@@ -1,51 +1,66 @@
-var   path = require('path')
+var   http = require('http')
+    , https = require('https')
+    , path = require('path')
     , konphyg = require('konphyg')(path.join(__dirname, '/config'))
     , express_cfg = konphyg('express')
     , lib = require(path.join(__dirname, '/lib'))
     , routes = require(path.join(__dirname, '/routes'))
-    //, consolidate = require('consolidate')
+    , consolidate = require('consolidate')
     , mongo_cfg = konphyg('mongo')
     , mongoose = require('mongoose')
     , db = mongoose.connect(mongo_cfg.db)
     , express = require('express')
-    , http = express.createServer()
-    , https = express.createServer(lib.authentication.options())
+    , app = express.createServer()
     , session = require('session-konphyg')
-    , connect_session = session.createSession();
+    , connect_session = session.createSession()
+    , fs = require('fs');
 
-lib.configure.server(http, connect_session);
-lib.configure.server(https, connect_session);
+app.engine('html', consolidate.dust);
+app.set('view engine', 'html');
+app.set('views', path.join(__dirname, express_cfg.view_path));
+app.set('db-uri', mongo_cfg.db);
+app.use(express.favicon());
+app.use(express.bodyParser());
+app.use(express.cookieParser('10001010101, this is top secret'));
+app.use(express.session(connect_session));
+app.use(express.methodOverride());
+app.use(express.static(path.join(__dirname, express_cfg.public_path_01)));
+app.use(express.static(path.join(__dirname, express_cfg.public_path_02)));
+app.use(app.router);
+app.use(lib.errors.invalid_password_handler);
+app.use(lib.errors.user_not_found_handler);
+app.use(lib.errors.user_not_authenticated_handler);
 
-//http routes
-http.get('/', routes.home.index);
-http.get('/home', routes.home.index);
-http.get('/about', routes.about.index);
-http.get('/contact', routes.contact.index);
-http.get('/github', routes.github.index);
-http.get('/login', routes.login.get_login);
-http.post('/login', routes.login.post_login);
-http.get('/account', lib.middleware.is_user_authenticated, routes.account.index);
-http.get('/logout', lib.middleware.is_user_authenticated, routes.login.logout);
-http.get('/404', lib.errors.page_not_found_handler);
-http.get('/500', lib.errors.internal_error_handler);
-http.post('/article', routes.article.save);
-http.get('/populate', routes.article.populate);
-
-//https routes
-https.post('/login', routes.login.post_login);
-https.get('/help', function(req, res){
-    res.json("HELP");
+//production settings
+app.configure('production', function() {
+    app.use(express.logger());
+    app.use(express.errorHandler());
 });
 
-//express listen
-var httpsServer = https.listen(443);
-var httpServer = http.listen(express_cfg.port);
-
-//socket.io setup
-var httpSocketIo = new require(path.join(__dirname, '/lib/socket_handler'))(httpServer, session.store(), session.options().sessionkey);
-var httpsSocketIo = new require(path.join(__dirname, '/lib/socket_handler'))(httpsServer, session.store(), session.options().sessionkey);
+//routes
+app.get('/', routes.home.index);
+app.get('/home', routes.home.index);
+app.get('/about', routes.about.index);
+app.get('/contact', routes.contact.index);
+app.get('/github', routes.github.index);
+app.get('/login', routes.login.get_login);
+app.post('/login', routes.login.post_login);
+app.get('/account', lib.middleware.is_user_authenticated, routes.account.index);
+app.get('/logout', lib.middleware.is_user_authenticated, routes.login.logout);
+app.get('/404', lib.errors.page_not_found_handler);
+app.get('/500', lib.errors.internal_error_handler);
+app.post('/article', routes.article.save);
+app.get('/populate', routes.article.populate);
 
 //handle uncaught exceptions
 process.addListener('uncaughtException', lib.errors.uncaught_exception);
 
-console.log('node plates - express on port %d in %s mode using session', express_cfg.port, http.settings.env, session.options().session_type);
+//http listen
+var httpServer = http.createServer(app.handle.bind(app)).listen(express_cfg.http_port);
+var httpSocketIo = new require(path.join(__dirname, '/lib/socket_handler'))(httpServer, session.store(), session.options().sessionkey);
+
+//https listen
+var httpsServer = https.createServer(lib.authentication.options(), app.handle.bind(app)).listen(express_cfg.https_port);
+var httpsSocketIo = new require(path.join(__dirname, '/lib/socket_handler'))(httpsServer, session.store(), session.options().sessionkey);
+
+console.log('node plates - express on port %d in %s mode using session', express_cfg.http_port, app.settings.env, session.options().session_type);